feat: 링크 삭제, 폴더 삭제 api 함수 생성 및 api 폴더 구조 세분화
